Moved OS to a folder to add plugs and other things
