+++ PoOS_OS/ProjectOrizonOS.sln
@@ -1,0 +1,24 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32112.339
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProjectOrizonOS", "ProjectOrizonOS.csproj", "{E933FA33-2311-4CB2-B224-DCC0EAC4B844}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E933FA33-2311-4CB2-B224-DCC0EAC4B844}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E933FA33-2311-4CB2-B224-DCC0EAC4B844}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E933FA33-2311-4CB2-B224-DCC0EAC4B844}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E933FA33-2311-4CB2-B224-DCC0EAC4B844}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {B91723FE-1C58-4FB6-877E-86F188FCD74C}
+	EndGlobalSection
+EndGlobal
