+++ PoOS_OS/Shell/Cmds/CommandManager.cs
