+++ PoOS_OS/Shell/Cmds/Console/cClear.cs
@@ -1,0 +1,11 @@
+﻿namespace ProjectOrizonOS.Shell.Cmds.Console
+{
+    internal class cClear
+    {
+        public static void Clear()
+        {
+            System.Console.Clear();
+        }
+    }
+    
+}
