+++ PoOS_OS/Shell/Cmds/Power/cReboot.cs
@@ -1,0 +1,13 @@
+﻿using Sys = Cosmos.System;
+
+namespace ProjectOrizonOS.Shell.Cmds.Power
+{
+    internal class cReboot
+    {
+        public static void Reboot()
+        {
+            System.Console.Clear();
+            Sys.Power.Reboot();
+        }
+    }
+}
