+++ PoOS_OS/Shell/Cmds/Power/cShutdown.cs
@@ -1,0 +1,13 @@
+﻿using Sys = Cosmos.System;
+
+namespace ProjectOrizonOS.Shell.Cmds.Power
+{
+    internal class cShutdown
+    {
+        public static void Shutdown()
+        {
+            System.Console.Clear();
+            Sys.Power.Shutdown();
+        }
+    }
+}
