--- ProjectOrizonOS.csproj
@@ -1,103 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
-        <!--<RuntimeIdentifier>cosmos</RuntimeIdentifier>-->
-        <Platform>cosmos</Platform>
-        <SupportsX86Intrinsics>false</SupportsX86Intrinsics>
-        <SelfContained>True</SelfContained>
-        <BinFormat>ELF</BinFormat>
-        <StackCorruptionDetectionEnabled>True</StackCorruptionDetectionEnabled>
-        <StackCorruptionDetectionLevel>MethodFooters</StackCorruptionDetectionLevel>
-        <Deployment>ISO</Deployment>
-        <DebugMode>Source</DebugMode>
-        <IgnoreDebugStubAttribute>False</IgnoreDebugStubAttribute>
-        <ISOFile>bin\Debug\net5.0\ProjectOrizonOS.iso</ISOFile>
-        <CompileVBEMultiboot>False</CompileVBEMultiboot>
-        <ExtractMapFile>True</ExtractMapFile>
-        <VMware_StackCorruptionDetectionEnabled>True</VMware_StackCorruptionDetectionEnabled>
-        <VMware_StackCorruptionDetectionLevel>MethodFooters</VMware_StackCorruptionDetectionLevel>
-        <VMware_Description>Use VMware Player or Workstation to deploy and debug.</VMware_Description>
-        <VMware_Deployment>ISO</VMware_Deployment>
-        <VMware_Launch>Bochs</VMware_Launch>
-        <VMware_DebugEnabled>True</VMware_DebugEnabled>
-        <VMware_DebugMode>Source</VMware_DebugMode>
-        <VMware_IgnoreDebugStubAttribute>False</VMware_IgnoreDebugStubAttribute>
-        <VMware_CosmosDebugPort>Serial: COM1</VMware_CosmosDebugPort>
-        <VMware_VisualStudioDebugPort>Pipe: Cosmos\Serial</VMware_VisualStudioDebugPort>
-        <VMware_PxeInterface>192.168.0.8</VMware_PxeInterface>
-        <VMware_VMwareEdition>Player</VMware_VMwareEdition>
-        <VMware_EnableGDB>False</VMware_EnableGDB>
-        <VMware_StartCosmosGDB>False</VMware_StartCosmosGDB>
-        <VMware_EnableBochsDebug>True</VMware_EnableBochsDebug>
-        <VMware_StartBochsDebugGui>False</VMware_StartBochsDebugGui>
-        <VMware_ISOFile>bin\Debug\net5.0\ProjectOrizonOS.iso</VMware_ISOFile>
-        <VMware_CompileVBEMultiboot>False</VMware_CompileVBEMultiboot>
-        <VMware_ExtractMapFile>True</VMware_ExtractMapFile>
-        <Bochs_StackCorruptionDetectionEnabled>True</Bochs_StackCorruptionDetectionEnabled>
-        <Bochs_StackCorruptionDetectionLevel>MethodFooters</Bochs_StackCorruptionDetectionLevel>
-        <Bochs_Description>Use Bochs emulator to deploy and debug.</Bochs_Description>
-        <Bochs_Deployment>ISO</Bochs_Deployment>
-        <Bochs_Launch>Bochs</Bochs_Launch>
-        <Bochs_DebugEnabled>True</Bochs_DebugEnabled>
-        <Bochs_DebugMode>Source</Bochs_DebugMode>
-        <Bochs_IgnoreDebugStubAttribute>False</Bochs_IgnoreDebugStubAttribute>
-        <Bochs_CosmosDebugPort>Serial: COM1</Bochs_CosmosDebugPort>
-        <Bochs_VisualStudioDebugPort>Pipe: Cosmos\Serial</Bochs_VisualStudioDebugPort>
-        <Bochs_PxeInterface>192.168.0.8</Bochs_PxeInterface>
-        <Bochs_VMwareEdition>Player</Bochs_VMwareEdition>
-        <Bochs_EnableGDB>False</Bochs_EnableGDB>
-        <Bochs_StartCosmosGDB>False</Bochs_StartCosmosGDB>
-        <Bochs_EnableBochsDebug>True</Bochs_EnableBochsDebug>
-        <Bochs_StartBochsDebugGui>False</Bochs_StartBochsDebugGui>
-        <Bochs_ISOFile>bin\Debug\net5.0\ProjectOrizonOS.iso</Bochs_ISOFile>
-        <Bochs_CompileVBEMultiboot>False</Bochs_CompileVBEMultiboot>
-        <Bochs_ExtractMapFile>True</Bochs_ExtractMapFile>
-    </PropertyGroup>
-
-    <PropertyGroup>
-        <EnableGDB>False</EnableGDB>
-        <StartCosmosGDB>False</StartCosmosGDB>
-        <VisualStudioDebugPort>Pipe: Cosmos\Serial</VisualStudioDebugPort>
-        <CosmosDebugPort>Serial: COM1</CosmosDebugPort>
-        <Launch>VMware</Launch>
-        <Profile>VMware</Profile>
-        <Description>Use VMware Player or Workstation to deploy and debug.</Description>
-        <PxeInterface>192.168.0.8</PxeInterface>
-        <DebugEnabled>True</DebugEnabled>
-        <VMWareEdition>Player</VMWareEdition>
-        <EnableBochsDebug>True</EnableBochsDebug>
-        <StartBochsDebugGui>False</StartBochsDebugGui>
-    </PropertyGroup>
-
-    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-      <Optimize>True</Optimize>
-    </PropertyGroup>
-
-    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-      <Optimize>False</Optimize>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <Compile Remove="Shell\Cmds\Network\**" />
-      <EmbeddedResource Remove="Shell\Cmds\Network\**" />
-      <None Remove="Shell\Cmds\Network\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="Shell\Cmds\Console\" />
-      <Folder Include="Shell\Graphics\" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Cosmos.Build" Version="0.1.0-localbuild20220204074457">
-        <PrivateAssets>all</PrivateAssets>
-        <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-      </PackageReference>
-      <PackageReference Include="Cosmos.Core" Version="0.1.0-localbuild20220204074509" />
-      <PackageReference Include="Cosmos.Debug.Kernel" Version="0.1.0-localbuild20220204074509" />
-      <PackageReference Include="Cosmos.System2" Version="0.1.0-localbuild20220204074509" />
-
-    </ItemGroup>
-
-</Project>
