--- Shell/Cmds/File/cRemoveFile.cs
@@ -1,23 +1,0 @@
-﻿using Cosmos.System.FileSystem.Listing;
-using Cosmos.System.FileSystem.VFS;
-using System;
-
-namespace ProjectOrizonOS.Shell.Cmds.File
-{
-    internal class cRemoveFile
-    {
-        private static ShellManager shell = new();
-
-        public static void RemoveFile(string file)
-        {
-            try
-            {
-                VFSManager.DeleteFile(Kernel.current_directory + file);
-            }
-            catch (Exception ex)
-            {
-                shell.WriteLine(ex.ToString(), type: 3);
-            }
-        }
-    }
-}
