--- Shell/Cmds/Power/cReboot.cs
@@ -1,13 +1,0 @@
-﻿using Sys = Cosmos.System;
-
-namespace ProjectOrizonOS.Shell.Cmds.Power
-{
-    internal class cReboot
-    {
-        public static void Reboot()
-        {
-            System.Console.Clear();
-            Sys.Power.Reboot();
-        }
-    }
-}
