--- Shell/Cmds/Power/cShutdown.cs
@@ -1,13 +1,0 @@
-﻿using Sys = Cosmos.System;
-
-namespace ProjectOrizonOS.Shell.Cmds.Power
-{
-    internal class cShutdown
-    {
-        public static void Shutdown()
-        {
-            System.Console.Clear();
-            Sys.Power.Shutdown();
-        }
-    }
-}
