--- Shell/ShellInfo.cs
@@ -1,11 +1,0 @@
-﻿namespace ProjectOrizonOS.Shell
-{
-    internal class ShellInfo
-    {
-        //Primary version Second version Revision BuildNumber
-        public static string version = "1.0.0.1";
-        public static string langSelected = "en_US";
-        public static bool firstRunning = true;
-        public static string user;
-    }
-}
